:recycle: refactor pinia persistence
