+++ src/stores/auth.js
@@ -1,51 +1,62 @@
 import { defineStore } from 'pinia'
 import { useApiCall } from '@/composables/network.js'
-import { useStorage } from '@vueuse/core'
+import { StorageSerializers, useStorage } from '@vueuse/core'
+import { computed } from 'vue'
 
-const authToken = useStorage('auth-token', null)
+export const useAuthStore = defineStore('auth', () => {
+  // States. We use localStorage to hydrate state when the page reloads
+  // @see https://github.com/vueuse/vueuse/pull/614#issuecomment-875450160 on why we need a serializer for `null` defaults
+  const authenticationToken = useStorage('auth-token', null, localStorage, { serializer: StorageSerializers.string })
+  const authenticatedUser = useStorage('auth-user', null, localStorage, { serializer: StorageSerializers.object })
 
-export const useAuthStore = defineStore('auth', {
-  state: () => ({
-    isAuthenticated: !!authToken.value,
-  }),
-  actions: {
-    login: async ({ email, password }) => {
-      const { data } = await useApiCall()('auth/tokens').post({
-        email,
-        password,
-        with_user: true,
-        client_name: `${navigator.platform} - Web`,
-      })
+  // Computed
+  const isAuthenticated = computed(() => {
+    return !!authenticationToken.value
+  })
 
-      const response = JSON.parse(data.value)
-      if (response.success) {
-        authToken.value = response.data.token
-      }
+  // Actions
+  const login = async ({ email, password }) => {
+    const { data } = await useApiCall()('auth/tokens').post({
+      email,
+      password,
+      with_user: true,
+      client_name: `${navigator.platform} - Web`,
+    })
 
-      return response
-    },
-    register: async ({ email, username, password, password_confirmation, first_name, last_name }) => {
-      const { data } = await useApiCall()('auth/register').post({
-        email,
-        username,
-        password,
-        password_confirmation,
-        first_name,
-        last_name,
-        client_name: `${navigator.platform} - Web`,
-      })
+    const response = JSON.parse(data.value)
+    if (response.success) {
+      authenticationToken.value = response.data.token
+      authenticatedUser.value = response.data.user
+    }
 
-      const response = JSON.parse(data.value)
-      if (response.success) {
-        authToken.value = response.data.token
-      }
+    return response
+  }
 
-      return response
-    },
-    checkAvailability: async (key, value) => {
-      if (!['username', 'email'].includes(key)) throw new Error('Key can only be "username" or "email"')
-      const { data } = await useApiCall()(`/availability/${key}?value=${value}`).get()
-      return JSON.parse(data.value)
-    },
-  },
+  const register = async ({ email, username, password, password_confirmation, first_name, last_name }) => {
+    const { data } = await useApiCall()('auth/register').post({
+      email,
+      username,
+      password,
+      password_confirmation,
+      first_name,
+      last_name,
+      client_name: `${navigator.platform} - Web`,
+    })
+
+    const response = JSON.parse(data.value)
+    if (response.success) {
+      authenticationToken.value = response.data.token
+      authenticatedUser.value = response.data.user
+    }
+
+    return response
+  }
+
+  const checkAvailability = async (key, value) => {
+    if (!['username', 'email'].includes(key)) throw new Error('Key can only be "username" or "email"')
+    const { data } = await useApiCall()(`/availability/${key}?value=${value}`).get()
+    return JSON.parse(data.value)
+  }
+
+  return { authenticationToken, authenticatedUser, isAuthenticated, login, register, checkAvailability }
 })
