+++ src/stores/theme.js
@@ -1,33 +1,27 @@
 import { defineStore } from 'pinia'
 import { useColorMode } from '@vueuse/core'
 
-const theme = useColorMode({
-  modes: {
-    // custom colors
-    dark: 'dark',
-    space: 'space',
-  },
-  emitAuto: true,
-})
-
-export const useThemeStore = defineStore('theme', {
-  state: () => ({
-    selectedTheme: theme.value,
-    availableThemes: [
-      {
-        key: 'dark',
-        name: 'Dark',
-      },
-      { key: 'light', name: 'Light' },
-      { key: 'space', name: 'Space' },
-      { key: 'old', name: 'Old School' },
-      { key: 'auto', name: 'System Preference' },
-    ],
-  }),
-  actions: {
-    switchTheme(selectedTheme) {
-      this.selectedTheme = selectedTheme
-      theme.value = selectedTheme
+export const useThemeStore = defineStore('theme', () => {
+  // States
+  const selectedTheme = useColorMode({
+    modes: { dark: 'dark', space: 'space' },
+    emitAuto: true,
+  })
+  const availableThemes = [
+    {
+      key: 'dark',
+      name: 'Dark',
     },
-  },
+    { key: 'light', name: 'Light' },
+    { key: 'space', name: 'Space' },
+    { key: 'old', name: 'Old School' },
+    { key: 'auto', name: 'System Preference' },
+  ]
+
+  // Actions
+  const switchTheme = (theme) => {
+    selectedTheme.value = theme
+  }
+
+  return { selectedTheme, availableThemes, switchTheme }
 })
