+++ src/views/misc/VerifyEmailGuardPage.vue
